requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.33.0
django-crispy-forms==1.6.1
django-djconfig==0.5.3
django-haystack==2.6.1
idna==2.6
mistune==0.7.4
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
Whoosh==2.7.4
